requirements.txt
sentence-transformers==2.2.2
datasets==2.11.0
numpy==1.24.2
